Halite custom TabCtrl working.
git-svn-id: svn://svn.geekisp.com/halite@113 362eace2-830a-0410-8e1a-3f32804a3aca

+++ src/DdxEx.hpp
@@ -34,7 +34,7 @@ class CWinDataExchangeEx : public CWinDataExchange<T>
 {
 public:	
 
-	BOOL DDX_StdWString(UINT id, wstring& str, BOOL save, const size_t max_len=MAX_PATH)
+	BOOL DDX_StdWString(UINT id, wstring& str, BOOL save, const int max_len=MAX_PATH)
 	{
 		T* pT = static_cast<T*>(this);
 		bool success = true;
